Extract auth middleware from service.
